+++ bookStory.ViewModels/Catalog/Translations/TranslationCreateRequest.cs
@@ -11,6 +11,6 @@ namespace bookStory.ViewModels.Catalog.Translations
         public Guid UserId { set; get; }
         public int IdParagraph { set; get; }
         public string Text { set; get; }
-        public int Rating { set; get; }
+        public float Rating { set; get; }
     }
 }
