+++ bookStory.ViewModels/Catalog/Translations/TranslationViewModel.cs
@@ -13,7 +13,7 @@ namespace bookStory.ViewModels.Catalog.Translations
         public Guid UserId { set; get; }
         public int IdParagraph { set; get; }
         public string Text { set; get; }
-        public int Rating { set; get; }
+        public float Rating { set; get; }
         public DateTime Date { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
