+++ bookStory.WebBookApp/Models/ParagraphDetailViewModel.cs
@@ -26,7 +26,7 @@ namespace bookStory.WebBookApp.Models
         public List<CommentViewModel> ListComments { get; set; }
         public int IdParagraph { set; get; }
         public string Text { set; get; }
-        public int Rating { set; get; }
+        public float Rating { set; get; }
         public int Vote { get; set; }
         public string Reason { set; get; }
         public string FirstName { get; set; }
